Fix bug where profileEnd() gets added even if no comment was found
This is due to a bug in logic. If we receive a `BlockStatement`, we were
adding `console.profileEnd()` even there was no `// profile` comment by
checking `!state.gotReturn` assuming this condition meant we reached the
end of the block but didn't get a return (`if` condition, for instance).
Since `gotReturn` is by default `false`, we were adding the closing
comment always. This introduces a perf issue in Safari because that
browser will try to run the profile by seeing this comment.